requirements.txt
flask
flask-cors
numpy
Pillow
requests
tensorflow
